+++ localflavor/ua/ua_regions.py
@@ -1,0 +1,33 @@
+from django.utils.translation import ugettext_lazy as _
+
+#: 24 oblasts, Avtonomna Respublika Krym and 2 cities with special status
+# Codes were gotten from ISO 3166-2:UA
+UA_REGION_CHOICES = (
+    ('UA-71', _('Cherkasy Oblast')),
+    ('UA-74', _('Chernihiv Oblast')),
+    ('UA-77', _('Chernivtsi Oblast')),
+    ('UA-12', _('Dnipropetrovsk Oblast')),
+    ('UA-14', _('Donetsk Oblast')),
+    ('UA-26', _('Ivano-Frankivsk Oblast')),
+    ('UA-63', _('Kharkiv Oblast')),
+    ('UA-65', _('Kherson Oblast')),
+    ('UA-68', _('Khmelnytskyi Oblast')),
+    ('UA-35', _('Kirovohrad Oblast')),
+    ('UA-32', _('Kiev Oblast')),
+    ('UA-09', _('Luhansk Oblast')),
+    ('UA-46', _('Lviv Oblast')),
+    ('UA-48', _('Mykolaiv Oblast')),
+    ('UA-51', _('Odessa Oblast')),
+    ('UA-53', _('Poltava Oblast')),
+    ('UA-56', _('Rivne Oblast')),
+    ('UA-59', _('Sumy Oblast')),
+    ('UA-61', _('Ternopil Oblast')),
+    ('UA-05', _('Vinnytsia Oblast')),
+    ('UA-07', _('Volyn Oblast')),
+    ('UA-21', _('Zakarpattia Oblast')),
+    ('UA-23', _('Zaporizhia Oblast')),
+    ('UA-18', _('Zhytomyr Oblast')),
+    ('UA-43', _('Autonomous Republic of Crimea')),
+    ('UA-30', _('Kiev')),
+    ('UA-40', _('Sevastopol'))
+)