+++ tests/test_ua/forms.py
@@ -1,0 +1,10 @@
+from django.forms import ModelForm
+
+from .models import UAFlavorTestModel
+
+
+class UAFlavorTestForm(ModelForm):
+
+    class Meta:
+        model = UAFlavorTestModel
+        fields = '__all__'